Fix broken semicolon merge
Merging https://github.com/MilliHealth/millimed/pull/1, master
was first merged into a different branch, and the information
from the semicolon merge was lost. This was a result of a
miscommunication about branch structuring. In the future,
to prevent such data loss, we're going to use a branch ->
master merge order, or we're going to just rebase branches
on top of master. The final option is to allow github to do
the heavy lifting and not worry about it at all ourselves.