Clean console errors/warnings
Before this commit, there are three of the same type
of errors regarding nesting <a> tags. This uses the
react-bootstrap as= property to use react router links,
which we need to have for refresh-less page switching,
but also have the Nav.Link styling.

The warning was around an unused variable in the
login actions module, which I just removed.