Fix a pernicious logout error:
So we were getting consistent errors on logout.
It turns out the reason for that was that **login** (aka authenticate)
was not respecting the Set-Cookie header, because 'withCredentials' was
not set to true within the axios config.

This fixes that issue, and with this fix it works for me (Aaron) in
firefox. However, I am still having an error in chromium receiving the
same 401. I suspect there's some sort of polluted Set-Cookie thing and
my browser is having trouble cleaning that pollution. On to continue
exploring that.