This prevents surveys from disappearing when finished
Our use case doesn't quite align with the Survey.js philosophy
of closing until refresh after the survey finsihes. I had an open
question (https://github.com/surveyjs/survey-library/issues/1855) which
was answered and fixes this issue.

However, there's also an issue, which is also in that question, about
_clearing_ the form when it's done, too, which currently does not work.
This fix _tries_ to, but it does not achieve that clearing. So another
question was asked in that same thread, and hopefully it comes with an
answer.

closes #26