Hide 'add new patient' option from patient picker
I'm not really sure how doctors are going to get patients to be with
them. Maybe it'd send an email invite to the patient, and the patient
would use that to both sign up and become a patient of that doctor? Sure
actually that sounds great but right now that functionality doesn't
exist so instead of popping up an alert when someone clicks it, we'll
just not give them the option to click it.

Admins can actually assign patients to doctors ATM.