Update patient picker behavior
I decided, after working with the app for a while, that as a doctor it's
a much nicer experience to stay on the same page when switching from
user to user. So I made that change, unless the doc is switching from no
user to a user, in which case it goes back to /overview.