Add provision for temporary signin help
Until we create the requisite accounts for signin, like sms sender and
email sender, we're just going to deliver the URL straight to the client
when they put in a phone number or email address. Note that this voids
all security, so this can't go live. It's for demo purposes only.